backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tweepy==4.14.0
openai==1.9.0
anthropic==0.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
numpy==1.26.3
celery==5.3.6
redis==5.0.1
structlog==24.1.0
prometheus_client==0.19.0
slowapi==0.1.9
python-dotenv==1.0.0
psycopg2-binary==2.9.9
